requirements.txt
django==1.11
Pillow==5.1.0
easy-thumbnails==2.5
django-image-cropping==1.1.0
django-ckeditor==5.4.0
django-phonenumber-field==2.0.0
django-widget-tweaks
django-cleanup
django-modeltranslation==0.13b1
django-rosetta==0.7.13
django-admin-sortable==2.1.5
pymemcache